ref(cmdk): Flatten modal.tsx into commandPalette.tsx
Move the CommandPaletteModal wrapper component and modalCss into
commandPalette.tsx, removing the separate modal.tsx indirection.
Update the dynamic import in actionCreators/modal.tsx to point to
the new location. Rename the spec file accordingly.
