+++ static/app/components/commandPalette/ui/commandPalette.modal.spec.tsx
@@ -25,8 +25,8 @@ import {
   CMDKAction,
   CommandPaletteProvider,
 } from 'sentry/components/commandPalette/ui/cmdk';
+import CommandPaletteModal from 'sentry/components/commandPalette/ui/commandPalette';
 import {CommandPaletteSlot} from 'sentry/components/commandPalette/ui/commandPaletteSlot';
-import CommandPaletteModal from 'sentry/components/commandPalette/ui/modal';
 import {
   makeCloseButton,
   makeClosableHeader,
